+++ zellij-utils/src/plugin_api/event.rs
@@ -2296,6 +2296,7 @@ fn serialize_mode_update_event_with_non_default_values() {
             // TODO: replace default
             rounded_corners: true,
             hide_session_name: false,
+            theme_owns_pane_defaults: false,
         },
         capabilities: PluginCapabilities { arrow_fonts: false },
         session_name: Some("my awesome test session".to_owned()),